+++ android/lib/grpc/src/main/kotlin/net/mullvad/mullvadvpn/lib/grpc/mapper/FromDomain.kt
@@ -217,6 +217,10 @@ internal fun ApiAccessMethod.fromDomain(): ManagementInterface.AccessMethod =
                     it.setEncryptedDnsProxy(
                         ManagementInterface.AccessMethod.EncryptedDnsProxy.getDefaultInstance()
                     )
+                is ApiAccessMethod.DomainFronting ->
+                    it.setDomainFronting(
+                        ManagementInterface.AccessMethod.DomainFronting.getDefaultInstance()
+                    )
             }
         }
         .build()