+++ ios/MullvadREST/ApiHandlers/RESTDefaults.swift
@@ -23,6 +23,9 @@ extension REST {
 
     public static let encryptedDNSHostname = infoDictionary["EncryptedDnsHostName"] as! String
 
+    public static let domainFrontingFront = infoDictionary["DomainFrontingFront"] as! String
+    public static let domainFrontingProxyHost = infoDictionary["DomainFrontingProxyHost"] as! String
+
     /// Disables API IP address cache when in staging environment and sticks to using default API endpoint instead.
     public static let isStagingEnvironment = false